+++ src/targets/gpu/device_name.cpp
@@ -87,6 +87,16 @@ bool gfx_default_rocblas()
 }
 #endif
 
+bool hipblaslt_supported()
+{
+    const auto device_name = trim(split_string(get_device_name(), ':').front());
+    // hipblaslt is supported for MI200 and above, and Navi3x and above.
+    return (device_name == "gfx90a" or
+            (starts_with(device_name, "gfx94") and device_name >= "gfx942") or
+            (starts_with(device_name, "gfx95") and device_name >= "gfx950") or
+            starts_with(device_name, "gfx110") or starts_with(device_name, "gfx120"));
+}
+
 } // namespace gpu
 } // namespace MIGRAPHX_INLINE_NS
 } // namespace migraphx
