+++ src/targets/gpu/hipblaslt.cpp
@@ -53,16 +53,6 @@ hipblaslt_preference_ptr create_hipblaslt_preference_ptr()
     return hipblaslt_preference_ptr{preference};
 }
 
-bool hipblaslt_supported()
-{
-    const auto device_name = trim(split_string(get_device_name(), ':').front());
-    // hipblaslt is supported for MI200 and above, and Navi3x and above.
-    return (device_name == "gfx90a" or
-            (starts_with(device_name, "gfx94") and device_name >= "gfx942") or
-            (starts_with(device_name, "gfx95") and device_name >= "gfx950") or
-            starts_with(device_name, "gfx110") or starts_with(device_name, "gfx120"));
-}
-
 #endif // MIGRAPHX_USE_HIPBLASLT
 
 } // namespace gpu
