+++ src/targets/gpu/lowering.cpp
@@ -33,6 +33,7 @@
 #include <migraphx/pass_manager.hpp>
 #include <migraphx/iterator_for.hpp>
 #include <migraphx/program.hpp>
+#include <migraphx/fp8_types.hpp>
 
 #include <migraphx/op/common.hpp>
 #include <migraphx/op/dot.hpp>
@@ -50,6 +51,7 @@
 #include <migraphx/gpu/rocblas.hpp>
 #include <migraphx/gpu/hipblaslt.hpp>
 #include <migraphx/gpu/compiler.hpp>
+
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
@@ -253,17 +255,32 @@ struct miopen_apply
             auto output = insert_allocation(ins, ins->get_shape());
             refs.push_back(output);
 
+#if not MIGRAPHX_USE_HIPBLASLT
+            return mod->replace_instruction(
+                ins, rocblas_gemm<Op>{Op{}, 1, 0, compute_fp32}, refs);
+#endif
+
 #if MIGRAPHX_USE_HIPBLASLT
+            bool has_fp8_inputs = std::any_of(ins->inputs().begin(),
+                    ins->inputs().end(),
+                    [](auto i_input){
+                        return contains(fp8_types{}.get(), i_input->get_shape().type());
+                    }
+            );
+
             // Check if user explicitly sets rocBLAS as GEMM provider, or
             // if the hardware cannot support hipblaslt, or
             // if the hardware is defaulted to use rocBLAS (such as gfx90).
-            if((string_value_of(MIGRAPHX_SET_GEMM_PROVIDER{}) == "rocblas") or
-               not hipblaslt_supported() or gpu::gfx_default_rocblas())
+            if (
+                not has_fp8_inputs and (
+                    (string_value_of(MIGRAPHX_SET_GEMM_PROVIDER{}) == "rocblas") or
+                    not hipblaslt_supported() or
+                    gpu::gfx_default_rocblas()
+                   )
+                )
             {
-#endif
                 return mod->replace_instruction(
                     ins, rocblas_gemm<Op>{Op{}, 1, 0, compute_fp32}, refs);
-#if MIGRAPHX_USE_HIPBLASLT
             }
             std::string op_name = "gpu::hip_gemm";
             if(contains(name, "quant_"))
